refactor: rename base component to common
