+++ src/components/task/create-new/index.jsx
@@ -2,9 +2,9 @@ import React, { useState } from "react";
 import PropTypes from "prop-types";
 import { useDispatch } from "react-redux";
 import DOMPurify from "dompurify";
-import Button from "components/base/button";
-import TextArea from "components/base/text-area";
-import Image from "components/base/image";
+import Button from "components/common/button";
+import TextArea from "components/common/text-area";
+import Image from "components/common/image";
 import { addTaskToTodo } from "store/actions/todo";
 import { generateTaskObject } from "utils/helper";
 import { ERROR_MESSAGE_TASK_TITLE } from "utils/const/errorMessages";
